requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cuid==0.4
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
libsql-client==0.2.2
multidict==6.0.4
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
yarl==1.9.2
